requirements.txt
Flask>=3.0,<4.0
Pillow>=10.3,<12.0
PyMuPDF>=1.24,<2.0
